[Mobile][UCC7][CA-108] Add ProductDetail

+++ lib/datasets/home_view.dart
@@ -54,8 +54,8 @@ class ProductListDataset {
         productInStoreId: json['productInStoreID'] as int,
         name: json['name'] as String,
         imagePath: json['imagePath'] as String,
-        oldPrice: (json['oldPrice'] as int).toDouble(),
-        salePrice: (json['salePrice'] as int).toDouble(),
+        oldPrice: json['oldPrice'].toDouble(),
+        salePrice: json['salePrice'].toDouble(),
         manufactureDate: DateTime.parse(json['manufactureDate'] as String),
         expireDate: DateTime.parse(json['expireDate'] as String));
   }
